requirements.txt
autopep8==1.4.4
dj-database-url==0.5.0
Django==2.2.5
django-appconf==1.0.3
django-compressor==2.3
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-libsass==0.7
django-sass-processor==0.7.3
gunicorn==19.9.0
libsass==0.19.2
pep8==1.7.1
Pillow==6.1.0
psycopg2==2.7.5
pycodestyle==2.5.0
pytz==2019.2
rcssmin==1.0.6
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.4
